style(clippy): collapse nested ifs and fix small lints
Resolve the 17 clippy errors blocking the build (deny(clippy::all)):
- Collapse nested if/match guards via if-let-chains (read_file,
  web_fetch, anthropic, gemini, minimax, openai, interactive).
- Replace manual modulo check with is_multiple_of (output).
- Drop redundant u16 cast (tui/app).
- Replace || std::io::stderr closures with std::io::stderr (mod).
- Restructure the empty-files branch of edit_file to avoid the
  unnecessary unwrap on files_value.

+++ src/cli/tui/app.rs
@@ -761,7 +761,7 @@ impl App {
         let _ = self.total_visual_rows(wrap_width);
         // +2 accounts for top and bottom borders of the completion Block widget.
         let completion_height: u16 = if has_completion {
-            ((self.cached_completion_rows + 2) as u16).min(output_area.height as u16)
+            ((self.cached_completion_rows + 2) as u16).min(output_area.height)
         } else {
             0
         };
